Suggestion: Add Clarification to Propulsion Types
This is a suggestion to add some additional clarification around propulsion types, mainly the difference between `electric_assist` and [`human`, `electric`]. I've seen a couple of provider APIs submitted that have the former when it should be the latter. I think that #48 did a great job of explaining why this should be represented as an array, but that reasoning is not reflected in the current MDS spec.