[Provider] Clarify how endpoints relate to municipality boundaries.
In conversations and from ingesting data, it's become clear there is
ambiguity in how parties are reporting data in their MDS Provider API
implementations with respect to the boundary of the municipality they
are operating in. This commit adds some clarifications and definitions
to capture the original intent of the spec and what most cities (Santa
Monica and LA) have been expecting in this regard.