Add new `battery_charged` event type reason
When the battery of a device is too low, the service yields the
`unavailable` event type with the `low_battery` reason. Then the
device may charge itself (for example because it has been plugged to a
charging dock by the last customer, or because the device has a solar
panel). Once the battery is charged enough, the service can yield an
`available` event type with the `battery_charged` reason.

[closes #210]