Flesh out definition of breaking vs. non-breaking changes
This commit adds a section to the release guidelines doc clarifying the
distinction between breaking and non-breaking changes, including an
enumeration of what kinds of changes fall in each category. This builds
on language added in #286; I didn't modify those templates here since
their description of breaking vs. non-breaking is still accurate.

I think the most likely area of controversy here is the handling of enum
values. My inclination would actually be to treat changes to enum values
as breaking changes, but it seems like the consensus during the 0.3.1
process so far has been that the benefits of having a lightweight
process for updating enums outweigh the risk of breaking clients that
are doing stricter validations.

Fixes #279