Update branching process
This commit modifies the branching mechanics in the release guidelines in
order to simplify the process of maintaining and contributing to MDS.

The most notable changes are:

* `dev` is now the target for almost all PRs, whether the changes they
contain are breaking or non-breaking. Maintainers are responsible for
backporting non-breaking changes from `dev` to the latest release
branch after merging the PR.

* Instead of merging changes into `master` and rebasing `dev`, `master`
is reset to point to the latest release as part of the release process.
Rebasing and merging between branches (other than PRs) are no longer
part of the process at all.

These changes are intended to be relatively lightweight and aimed at
smoothing out pain points with the current workflow; at some point we
might want to consider [more significant changes] like organizing the
overall directory structure by version.

[more significant changes]: https://github.com/CityOfLosAngeles/mobility-data-specification/issues/288#issuecomment-480360455