Clarify and simplify geometry-driven events by removing trip_leave_geography event
The inclusion of the leave event introduced complexity and ambiguity around which event_geographies should be communicated. For every other event, event_geographies simply consists of every Geography that contains the location of the event. However, for a leave event it is unclear whether event_geographies should contain the geographies the vehicle just left. Removing this event resolves this ambiguity, the list of geographies left can be inferred if neeeded.