Clarify Provider API responses
The trips and status changes endpoints are expected to return
HTTP 404 responses when data "does not exist". I'm trying to
clarify that that means the future and hours when the provider
was not in operation in the municipality. For past hours during
which the provider was operating but nothing happened they should
return "200 OK" responses with  empty arrays of trips or status
changes. Additionally, for hours that are not available in the API
because the data is still be processed the API shall return a
"102 Processing" response to indicate the requester should check back
later.