Allow users to specify rule min/max treatment
MDS does not currently specify whether Policy Rule minimum
and maximum values are considered within bounds when
evaluating whether something matches the rule
(in other words, whether the rule bounds are a open, closed,
or half-open interval). This PR adds new fields to Policy Rules
that allow users to individually specify the treatment of the
minimum and maximum values.

The new fields are called inclusive_minimum and inclusive_maximum.
They are boolean fields that both default to true.