Remove word data from response payload descriptions
In older versions of MDS data was nested under a `data`
key in provider API responses, so trips data would e.g.
be response['data']['trips']. This ended up reflected
in the docs where each section would have a label like
"`data` Payload" describing what was under the `data`
key in the payload.

In MDS 2.0 this nesting was removed, `trips` is now a top
level key in the response from provider trips endpoint.
However, the docs were not updated to remove the word
"data" from the payload descriptions, leading to potential
confusion for readers of the documentation. This PR
changes "`data` Payload" to "Payload" throughout the provider
README to bring it in line with the spec.