Update provider vehicles endpoints descriptions
The description of the Provider vehicles endpoints was never
really updated in the MDS 1 -> 2 transition so this rearranges
and clarifies some things now that there are two vehicle information
endpoints, one for static information like vehicle/propulsion type
and another for realtime status. These changes are mostly
rearranging/clarifying, but there is one major addition to the
description of the `/vehicles` endpoint describing that when
called without a device ID it is expected to return every vehicle
ever deployed in a jurisdiction.